Revert "ANDROID: binder: fix format specifier for type binder_size_t"
This reverts commit df3087d4836dd35c138c628c50b0986169183ee9.

+++ drivers/android/binder.c
@@ -2066,7 +2066,7 @@ static void binder_transaction(struct binder_proc *proc,
 	if (!IS_ALIGNED(extra_buffers_size, sizeof(u64))) {
 		binder_user_error("%d:%d got transaction with unaligned buffers size, %llu\n",
 				  proc->pid, thread->pid,
-				  (u64)extra_buffers_size);
+				  extra_buffers_size);
 		return_error = BR_FAILED_REPLY;
 		goto err_bad_offset;
 	}
